+++ NEWS.md
@@ -5,6 +5,7 @@ Version: development
 --------------------
 * The get_max_value/get_min_value use torch.max/min when no missing data is present in the column
 * Fixed upgrade_frame bug
+* Fixed bug with saving CSV-files when they contain boolean values
 
 Version: 1.6
 --------------------